Split coordinator and sensor
